Increase timeout for stop callSync
:Release Notes:
Increased the timeout for the callSync function to 16000 ms

:Detailed Notes:
The callSync function in media_recorder.cpp was updated to include
a timeout of 16000 milliseconds. This change aims to improve the
reliability of synchronous calls by ensuring sufficient time is
allocated for the operation to complete, reducing the chance of
timeout errors in the recording process.

:Testing Performed:
local build & test OK

:QA Notes:
N/A

:Issues Addressed:
[WRR-16913] [ACP Reference] API TC Preparation - Media Recorder Service

Change-Id: I225224c4ef201f5e5de68e97435c15ecea7a3a0f
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.service.mediarecorder/+/465465
Reviewed-by: AI Reviewer <ai_reviewer@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>
Tested-by: Sungho Lee <shl.lee@lge.com>
Reviewed-by: Sungho Lee <shl.lee@lge.com>